dashboard/pyproject.toml
[tool.poetry]
name = "7apps-dashboard"
version = "0.1.0"
description = "Landing page and app dashboard for the 7Apps demo."
authors = ["Chris Tippett <chris.tippett@hey.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.60.1"
google-cloud-pubsub = "^1.7.0"
google-auth = "^1.19.2"
python-dateutil = "^2.8.1"
python-dotenv = "^0.14.0"
jinja2 = "^2.11.2"
aiofiles = "^0.5.0"
uvicorn = "^0.11.6"
aiohttp = "^3.6.2"
PyYAML = "^5.3.1"

[tool.poetry.dev-dependencies]
pytest = "^5.4.3"
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
